refactor(cli): fold resume candidates under resume
Problem: #1842 still exposed resume-candidates as a separate root even though it is part of the resume/context workflow.

What changed: make resume a real lazy Click group, keep polylogue resume SESSION_ID through a hidden default session command, move the candidate-ranking command to polylogue resume candidates, delete the old resume_candidates module, and refresh generated CLI docs/help/topology snapshots.

Verification: devtools test tests/unit/cli/test_resume.py tests/unit/cli/test_json_output.py tests/unit/cli/test_click_app.py::TestCliMetadata::test_retained_command_owners_are_registered tests/unit/cli/test_help_snapshots.py tests/unit/cli/test_terminal_snapshots.py tests/unit/cli/test_plain_cli_snapshots.py tests/unit/showcase/test_exercise_catalog.py::TestExercisesByGroup::test_all_command_paths_have_generated_help_exercises tests/unit/showcase/test_exercise_catalog.py::TestExercisesByGroup::test_inventory_includes_nested_command_paths -> 87 passed; devtools lab scenario verify-baselines -> 96 exercises, all baselines match; devtools render all --check -> sync OK; devtools verify --quick -> exit_code 0.

Ref #1842
