+++ docs/topology-status.md
@@ -28,10 +28,10 @@ Generated by `devtools render topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 700
+- **Stable** (no move scoped): 699
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 841
-- **Realized polylogue/**/*.py**: 841 files declared
+- **Total declared**: 840
+- **Realized polylogue/**/*.py**: 840 files declared
 <!-- END GENERATED: topology-status -->
