+++ tests/baselines/showcase/help-resume.txt
@@ -1,8 +1,9 @@
-Usage: polylogue resume [OPTIONS] SESSION_ID
+Usage: polylogue resume [OPTIONS] COMMAND [ARGS]...
 
   Reconstruct work-state context for a fresh agent session.
 
 Options:
-  --related-limit INTEGER  Maximum related sessions to include.  [default: 6]
-  -f, --format [json]      Output format.
-  -h, --help               Show this message and exit.
+  -h, --help  Show this message and exit.
+
+Commands:
+  candidates  Rank archived logical sessions for the current working...
